grap_altimeter/scan: refactor postprocessing code
